a few ioquake3 fixes

+++ xreal/code/q3_ui/ui_connect.c
@@ -120,7 +120,7 @@ static void UI_DisplayDownloadInfo(const char *downloadName)
 
 	if(downloadSize > 0)
 	{
-		s = va("%s (%d%%)", downloadName, downloadCount * 100 / downloadSize);
+		s = va("%s (%d%%)", downloadName, (int)((float)downloadCount * 100.0f / downloadSize));
 	}
 	else
 	{
